+++ packages/validator-ajv8/src/createAjvInstance.ts
@@ -1,12 +1,17 @@
-import Ajv, { Options } from "ajv";
+import Ajv, { Options } from "ajv8";
 import addFormats, { FormatsPluginOptions } from "ajv-formats";
 import isObject from "lodash/isObject";
 
 import { CustomValidatorOptionsType } from "./types";
+import {
+  ADDITIONAL_PROPERTY_FLAG,
+  RJSF_ADDITONAL_PROPERTIES_FLAG,
+} from "@rjsf/utils";
 
 export const AJV_CONFIG: Options = {
   allErrors: true,
   multipleOfPrecision: 8,
+  strict: false,
 } as const;
 export const COLOR_FORMAT_REGEX =
   /^(#?([0-9A-Fa-f]{3}){1,2}\b|aqua|black|blue|fuchsia|gray|green|lime|maroon|navy|olive|orange|purple|red|silver|teal|white|yellow|(rgb\(\s*\b([0-9]|[1-9][0-9]|1[0-9][0-9]|2[0-4][0-9]|25[0-5])\b\s*,\s*\b([0-9]|[1-9][0-9]|1[0-9][0-9]|2[0-4][0-9]|25[0-5])\b\s*,\s*\b([0-9]|[1-9][0-9]|1[0-9][0-9]|2[0-4][0-9]|25[0-5])\b\s*\))|(rgb\(\s*(\d?\d%|100%)+\s*,\s*(\d?\d%|100%)+\s*,\s*(\d?\d%|100%)+\s*\)))$/;
@@ -26,6 +31,7 @@ export const DATA_URL_FORMAT_REGEX =
  * @param [customFormats] - The set of additional custom formats that the validator will support
  * @param [ajvOptionsOverrides={}] - The set of validator config override options
  * @param [ajvFormatOptions] - The `ajv-format` options to use when adding formats to `ajv`; pass `false` to disable it
+ * @param [AjvClass] - The `Ajv` class to use when creating the validator instance
  */
 export default function createAjvInstance(
   additionalMetaSchemas?: CustomValidatorOptionsType["additionalMetaSchemas"],
@@ -35,13 +41,19 @@ export default function createAjvInstance(
   AjvClass: typeof Ajv = Ajv
 ) {
   const ajv = new AjvClass({ ...AJV_CONFIG, ...ajvOptionsOverrides });
-  if (typeof ajvFormatOptions !== "boolean") {
+  if (ajvFormatOptions) {
     addFormats(ajv, ajvFormatOptions);
+  } else if (ajvFormatOptions !== false) {
+    addFormats(ajv);
   }
 
   // add custom formats
   ajv.addFormat("data-url", DATA_URL_FORMAT_REGEX);
   ajv.addFormat("color", COLOR_FORMAT_REGEX);
+
+  // Add RJSF-specific additional properties keywords so Ajv doesn't report errors if strict is enabled.
+  ajv.addKeyword(ADDITIONAL_PROPERTY_FLAG);
+  ajv.addKeyword(RJSF_ADDITONAL_PROPERTIES_FLAG);
 
   // add more schemas to validate against
   if (Array.isArray(additionalMetaSchemas)) {