+++ packages/validator-ajv8/test/createAjvInstance.test.ts
@@ -1,5 +1,5 @@
-import Ajv from "ajv";
-import Ajv2019 from "ajv/dist/2019";
+import Ajv from "ajv8";
+import Ajv2019 from "ajv8/dist/2019";
 import addFormats from "ajv-formats";
 
 import createAjvInstance, {
@@ -9,8 +9,8 @@ import createAjvInstance, {
 } from "../src/createAjvInstance";
 import { CustomValidatorOptionsType } from "../src";
 
-jest.mock("ajv");
-jest.mock("ajv/dist/2019");
+jest.mock("ajv8");
+jest.mock("ajv8/dist/2019");
 jest.mock("ajv-formats");
 
 export const CUSTOM_OPTIONS: CustomValidatorOptionsType = {
@@ -42,7 +42,7 @@ describe("createAjvInstance()", () => {
       expect(Ajv).toHaveBeenCalledWith(AJV_CONFIG);
     });
     it("expect addFormats to be called with the new ajv instance and undefined", () => {
-      expect(addFormats).toHaveBeenCalledWith(ajv, undefined);
+      expect(addFormats).toHaveBeenCalledWith(ajv);
     });
     it("addFormat() was called twice", () => {
       expect(ajv.addFormat).toHaveBeenCalledTimes(2);
@@ -87,7 +87,7 @@ describe("createAjvInstance()", () => {
       expect(Ajv).not.toHaveBeenCalled();
     });
     it("expect addFormats to be called with the new ajv instance and undefined", () => {
-      expect(addFormats).toHaveBeenCalledWith(ajv, undefined);
+      expect(addFormats).toHaveBeenCalledWith(ajv);
     });
     it("addFormat() was called twice", () => {
       expect(ajv.addFormat).toHaveBeenCalledTimes(2);